added component folders, and routes file
